# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Live Score — Housing Trust Infrastructure API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List all partners
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Partner list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a partner (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Created partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/maintenance/cleanup-logos:
    post:
      operationId: cleanupPartnerLogos
      tags: [partners]
      summary: Review or remove orphaned partner logo objects (admin)
      description: Lists managed partner logo objects that are no longer referenced by a partner. The default is a dry-run; set dryRun to false to remove the reviewed orphaned objects.
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerLogoCleanupInput"
      responses:
        "200":
          description: Logo cleanup review result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerLogoCleanupResult"
        "400":
          description: Invalid cleanup request
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
  /partners/maintenance/cleanup-logos/history:
    get:
      operationId: listPartnerLogoCleanupHistory
      tags: [partners]
      summary: List prior partner logo cleanup runs (admin)
      description: Returns the immutable history of logo cleanup dry-runs and confirmed runs. This endpoint is read-only.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          description: Filter cleanup history by whether any deletion failed.
          schema:
            type: string
            enum: [all, failed]
            default: all
          required: false
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
          required: false
      responses:
        "200":
          description: Prior logo cleanup runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerLogoCleanupHistoryEntry"
        "400":
          description: Invalid history query
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned object storage upload URL
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageUploadRequest"
      responses:
        "200":
          description: Presigned upload URL and object path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUploadResponse"
  /storage/kyc/uploads/request-url:
    post:
      operationId: requestKycUploadUrl
      tags: [users]
      summary: Request a private upload URL for a verification document
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycUploadRequest"
      responses:
        "200":
          description: Private upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUploadResponse"
        "400":
          description: Unsupported file
        "401":
          description: Authentication required
        "503":
          description: Object storage unavailable
...
  /partners/{id}:
    get:
      operationId: getPartner
      tags: [partners]
      summary: Get partner by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update a partner (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerUpdate"
      responses:
        "200":
          description: Updated partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    delete:
      operationId: deletePartner
      tags: [partners]
      summary: Delete a partner (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: Deleted
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    KycUploadRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          maxLength: 160
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        contentType:
          type: string
          enum: [application/pdf, image/jpeg, image/png, image/webp]
...
    Partner:
      type: object
      required: [id, name, category, bonusMultiplier, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [rent, utilities, transport, health, food, other]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        logoObjectPath:
          type: ["string", "null"]
        bonusMultiplier:
          type: number
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        isActive:
          type: boolean
        totalTransactions:
          type: ["integer", "null"]
        totalVolumeRub:
          type: ["number", "null"]
    PartnerLogoCleanupInput:
      type: object
      properties:
        dryRun:
          type: boolean
          default: true
    PartnerLogoCleanupFailedPath:
      type: object
      required: [path, error]
      properties:
        path:
          type: string
        error:
          type: string
    PartnerLogoCleanupResult:
      type: object
      required: [dryRun, scanned, referenced, orphaned, removed, failed]
      properties:
        dryRun:
          type: boolean
        scanned:
          type: integer
        referenced:
          type: integer
        orphaned:
          type: array
          items:
            type: string
        removed:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            $ref: "#/components/schemas/PartnerLogoCleanupFailedPath"
    PartnerLogoCleanupHistoryEntry:
      type: object
      required: [id, adminUserId, adminName, adminPhone, mode, scanned, referenced, orphaned, removed, failed, createdAt]
      properties:
        id:
          type: integer
        adminUserId:
          type: integer
        adminName:
          type: string
        adminPhone:
          type: ["string", "null"]
        mode:
          type: string
          enum: [dry_run, confirmed]
        scanned:
          type: integer
        referenced:
          type: integer
        orphaned:
          type: array
          items:
            type: string
        removed:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            $ref: "#/components/schemas/PartnerLogoCleanupFailedPath"
        createdAt:
          type: string
    StorageUploadRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
    StorageUploadResponse:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          required: [name, size, contentType]
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
    PartnerInput:
      type: object
      required: [name, category, bonusMultiplier]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [rent, utilities, transport, health, food, other]
        description:
          type: string
        logoUrl:
          type: string
        logoObjectPath:
          type: string
        bonusMultiplier:
          type: number
        address:
          type: string
        city:
          type: string
    PartnerUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [rent, utilities, transport, health, food, other]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        logoObjectPath:
          type: ["string", "null"]
        bonusMultiplier:
          type: number
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
...